vf1

+++ app/Http/Controllers/AgendamentoApiController.php
@@ -3,10 +3,16 @@
 namespace App\Http\Controllers;
 
 use App\FIlters\AgendamentoFIlter;
-use App\Http\Controllers\Resource\AgendamentoResource;
+use App\Http\Requests\AgendamentoRequest;
+use App\Http\Resources\AgendamentoResource;
 use App\Repository\AgendamentosRepository;
+
+
 use App\Repository\UserRepository;
 use Illuminate\Http\Request;
+use App\Models\Agendamento;
+use Illuminate\Http\JsonResponse;
+
 
 class AgendamentoApiController extends Controller
 {
@@ -21,11 +27,12 @@ class AgendamentoApiController extends Controller
     //GET /api/eventos - Listar todos os eventos - INDEX => convenção
     public function index(Request $request)
     {
+        $this->authorize('viewAny', Agendamento::class);
         try {
 
             $filter = new AgendamentoFIlter($request);
             $perpage = $request->input('perpage', 15);
-            $getAgendamento = $this->repository->filterPaginate($filter, $perpage);
+            $getAgendamento = $this->repository->filterPaginated($filter, $perpage);
             return AgendamentoResource::collection($getAgendamento);
         } catch (\Exception $ex) {
             return response()->json([
@@ -36,11 +43,13 @@ class AgendamentoApiController extends Controller
     }
 
 
-    public function store(Request $request)
+    public function store(AgendamentoRequest $request): JsonResponse
     {
+        $this->authorize('create', Agendamento::class);
         try {
 
-            $dados = $request->all();
+            $dados = $request->validated();
+            $dados['status'] = $dados['status'] ?? 'confirmado';
             $salvar = $this->repository->salvar($dados);
             return response()->json([
                 'message' =>  'local criado com sucesso',
@@ -60,11 +69,12 @@ class AgendamentoApiController extends Controller
     /**
      * Display the resource.
      */
-    public function show($id)
+    public function show($id): JsonResponse
     {
         try {
             $get =  $this->repository->getId($id);
-            return new AgendamentoResource($get);
+            $this->authorize('view', $get );
+            return response()->json([new AgendamentoResource($get)]);
         } catch (\Exception $ex) {
             return response()
                 ->json([
@@ -73,21 +83,21 @@ class AgendamentoApiController extends Controller
         }
     }
 
-    /**
-     * Update the resource in storage.
-     */
-    public function update(Request $request, $id)
+
+    public function update(AgendamentoRequest $request, $id): JsonResponse
     {
         $id = $this->repository->getId($id);
+        if (!$id) {
+            return response()->json([
+                "message" => 'o id não pôde ser localizado'
+            ], 404);
+        }
+        $this->authorize('update', $id);
         try {
 
-            if (!$id) {
-                return response()->json([
-                    "message" => 'o id não pôde ser localizado'
-                ], 404);
-            }
-            $local = $request->all();
-            $local['id'] = $id;
+            //array merge pega o array que está no request e junta id => $id fazendo um array só
+            $local =  array_merge($request->validated(), ['id' => $id]);
+
             $update = $this->repository->salvar($local);
             return response()->json([
                 "message" => 'evento atualizado com sucesso',
@@ -100,44 +110,32 @@ class AgendamentoApiController extends Controller
         }
     }
 
-    /**
-     * Remove the resource from storage.
-     */
-    public function destroy($id) {
+
+    public function destroy($id)
+    {
 
 
         $agendamentoId = $this->repository->getId($id);
+        if (!$agendamentoId) {
+            return response()->json([
+                "message" => 'usuario não encontrado'
+            ], 404);
+        }
+        $this->authorize('delete', $agendamentoId);
+        try {
 
-            try {
-
-                if (!$agendamentoId) {
-                    return response()->json([
-                        "message" => 'usuario não encontrado'
-                    ], 404);
-                } else{
-
-                    $userId = auth('sanctum')->id();
-                    if($agendamentoId->user_id !== $userId ){
-                        return response()->json(["message" => 'usuario não tem permissão para deletar esse agendamento'
-                    ], 403);
-                    }
             $this->repository->destroyId($agendamentoId);
-                return response()->json(['message' => 'local deletado com sucesso',
+            return response()->json([
+                'message' => 'local deletado com sucesso',
 
-            ], 200);}
+            ], 200);
+        } catch (\Exception $ex) {
 
+            return response()->json([
+                'message' => 'não foi possível deletar o registro',
+                'erro' => $ex->getMessage(),
 
-
-}catch (\Exception $ex){
-
-    return response()->json(['message' => 'não foi possível deletar o registro',
-        'erro' => $ex->getMessage(),
-
-], 500);
-}
-
-}
-
-
-
+            ], 500);
+        }
+    }
 }
